requirements.txt
flask>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
openai>=1.12.0
sentence-transformers>=2.5.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
Pillow>=10.0.0
transformers>=4.30.0
torch>=2.0.0